Improved network finished, need to test against the MINST dataset of handwritten digets
